move dataframe_helper to fa folder since only used there
