+++ gamestonk_terminal/fundamental_analysis/alpha_vantage_view.py
@@ -9,7 +9,7 @@ from alpha_vantage.fundamentaldata import FundamentalData
 import pandas as pd
 
 from gamestonk_terminal import config_terminal as cfg
-from gamestonk_terminal.dataframe_helpers import clean_df_index
+from gamestonk_terminal.fundamental_analysis.fa_helper import clean_df_index
 from gamestonk_terminal.helper_funcs import (
     check_positive,
     long_number_format,
